AllTeacher/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
gunicorn==22.0.0
openai==1.54.4
supabase==2.8.1
pyjwt[crypto]==2.9.0
requests==2.32.3
